backend/utils/calendar_utils.py
"""
Meeting Ghost AI — Calendar Utilities
=======================================
Google Calendar API integration for meeting detection and auto-join triggering.
"""

import logging
from datetime import datetime, timedelta
from typing import Optional
from config import settings

logger = logging.getLogger(__name__)


class CalendarService:
    """Google Calendar integration for meeting detection."""

    def __init__(self):
        self.client = None
        self._init_client()

    def _init_client(self):
        """Initialize Google Calendar API client."""
        if not settings.google_calendar_client_id:
            logger.info("Google Calendar not configured. Manual meeting creation only.")
            return
        try:
            from google.oauth2.credentials import Credentials
            from googleapiclient.discovery import build
            # In production, use proper OAuth flow
            logger.info("Google Calendar client initialized.")
        except ImportError:
            logger.warning("Google API libraries not installed.")

    async def get_upcoming_meetings(self, minutes_ahead: int = 30) -> list[dict]:
        """Get meetings starting in the next N minutes."""
        if not self.client:
            return []
        try:
            now = datetime.utcnow()
            time_max = now + timedelta(minutes=minutes_ahead)
            # Calendar API call would go here
            return []
        except Exception as e:
            logger.error(f"Calendar fetch error: {e}")
            return []

    async def get_meeting_details(self, event_id: str) -> Optional[dict]:
        """Get details for a specific calendar event."""
        if not self.client:
            return None
        return None

    def create_mock_meetings(self) -> list[dict]:
        """Create mock meetings for development."""
        now = datetime.utcnow()
        return [
            {"id": "mock-1", "title": "Sprint Planning", "start": now.isoformat(),
             "end": (now + timedelta(hours=1)).isoformat(),
             "link": "https://meet.google.com/mock-meeting", "attendees": ["Manager", "Dev Team"]},
            {"id": "mock-2", "title": "Design Review", "start": (now + timedelta(hours=2)).isoformat(),
             "end": (now + timedelta(hours=3)).isoformat(),
             "link": "https://zoom.us/j/mock-meeting", "attendees": ["Designer", "Product"]},
        ]
