requirements.txt
# ============================================
# Web Framework
# ============================================
fastapi==0.109.2
uvicorn[standard]==0.27.1
websockets==12.0
python-multipart==0.0.9

# ============================================
# AWS SDK
# ============================================
boto3==1.34.49
botocore==1.34.49

# ============================================
# AI / ML  (ORDER MATTERS: torch must install before sentence-transformers)
# ============================================
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers==2.3.1
deepgram-sdk>=3.0.0

# ============================================
# Vector Database
# ============================================
pinecone-client==3.0.2

# ============================================
# Database
# ============================================
sqlalchemy==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9

# ============================================
# Caching
# ============================================
redis==5.0.1

# ============================================
# Audio Processing
# ============================================
soundfile==0.12.1
numpy==1.26.4
scipy==1.12.0
pydub==0.25.1

# ============================================
# Utilities
# ============================================
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.27.0
aiofiles==23.2.1

# ============================================
# Google Calendar
# ============================================
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0

# ============================================
# Development
# ============================================
pytest==7.4.4
pytest-asyncio==0.23.4
black==24.2.0
ruff==0.2.1

# ============================================
# Speech to Text
# ============================================
openai-whisper==20231117
ffmpeg-python==0.2.0
